refactor: fold signerToGuid from guid.ts into shared.ts
Small single-function module absorbed into existing utilities file.
