+++ packages/controller/src/session/ts/index.ts
@@ -1,6 +1,5 @@
 export * from "./types";
 export * from "./errors";
-export * from "./guid";
 export * from "./merkle";
 export * from "./execution";
 export * from "./account";
