+++ packages/controller/src/session/ts/shared.ts
@@ -1,4 +1,12 @@
-import { addAddressPadding, hash, num } from "starknet";
+import {
+  addAddressPadding,
+  ec,
+  encode,
+  hash,
+  num,
+  shortString,
+} from "starknet";
+import type { Signer } from "./types";
 
 /**
  * Normalize a felt value to a lowercase hex string.
@@ -29,4 +37,26 @@ export function normalizeContractAddress(
     throw new Error(`${context} is missing a contract address.`);
   }
   return addAddressPadding(trimmed.toLowerCase());
+}
+
+const STARKNET_SIGNER_DOMAIN = num
+  .toHex(shortString.encodeShortString("Starknet Signer"))
+  .toLowerCase();
+
+/**
+ * Derives a session signer GUID from a Signer object.
+ * Matches the WASM `signerToGuid` function signature exactly.
+ */
+export function signerToGuid(signer: Signer): string {
+  const privateKey = signer.starknet?.privateKey;
+  if (!privateKey) {
+    throw new Error("Cannot derive session GUID: missing starknet private key");
+  }
+
+  const normalizedKey = encode.addHexPrefix(String(privateKey).trim());
+  const publicKey = ec.starkCurve.getStarkKey(normalizedKey);
+
+  return num
+    .toHex(hash.computePoseidonHash(STARKNET_SIGNER_DOMAIN, publicKey))
+    .toLowerCase();
 }
